Clean up global variables somewhat
The extemely lengthy lists of 'extern this and that' at the top of
every file was error prone and hard to maintain. Having the variable
definitions stored in base/nagios.c also made it nigh-impossible to
link the sources to another main() routine without creating several
stub variables that were never used.

The largest common denominator for core-only variables seems to be
base/utils.c (since that's where reset_variables() lives), so we
move the definitions there, and put core-only declarations in
include/nagios.h. Variables shared between core and cgi's are
defined in common/shared.c and declared in include/common.h.

The same work could be done for cgi's, but the savings would be
considerably smaller there, so I refrained.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2168 5f96b256-904b-4d8d-8c98-d829582c6739

+++ base/nagiostats.c
@@ -46,7 +46,7 @@ char *status_version = NULL;
 time_t program_start = 0L;
 int status_service_entries = 0;
 int status_host_entries = 0;
-unsigned long nagios_pid = 0L;
+int nagios_pid = 0;
 
 double min_service_state_change = 0.0;
 int have_min_service_state_change = FALSE;
@@ -454,9 +454,9 @@ int display_mrtg_values(void) {
 		else if(!strcmp(temp_ptr, "NAGIOSVERSION"))
 			printf("%s%s", status_version, mrtg_delimiter);
 		else if(!strcmp(temp_ptr, "NAGIOSPID"))
-			printf("%lu%s", nagios_pid, mrtg_delimiter);
+			printf("%d%s", nagios_pid, mrtg_delimiter);
 		else if(!strcmp(temp_ptr, "NAGIOSVERPID"))
-			printf("Nagios %s (pid=%lu)%s", status_version, nagios_pid, mrtg_delimiter);
+			printf("Nagios %s (pid=%d)%s", status_version, nagios_pid, mrtg_delimiter);
 
 
 		else if(!strcmp(temp_ptr, "NUMSERVICES"))
@@ -759,7 +759,7 @@ int display_stats(void) {
 	time_difference = (current_time - program_start);
 	get_time_breakdown(time_difference, &days, &hours, &minutes, &seconds);
 	printf("Program Running Time:                   %dd %dh %dm %ds\n", days, hours, minutes, seconds);
-	printf("Nagios PID:                             %lu\n", nagios_pid);
+	printf("Nagios PID:                             %d\n", nagios_pid);
 	printf("\n");
 	printf("Total Services:                         %d\n", status_service_entries);
 	printf("Services Checked:                       %d\n", services_checked);
